Move CursorData into Geometry namespace
