+++ Ciallo/Rendering/WorldCursorDetectionArea.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using Ciallo.Data;
-using Ciallo.NodeControl;
+using Ciallo.Geometry;
 using Godot;
 using Godot.Collections;
 using R3;
