+++ Ciallo/Tool/DragStrokeInteractor.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-using Ciallo.NodeControl;
+using Ciallo.Geometry;
 
 namespace Ciallo.Tool;
