+++ Ciallo/Tool/InteractorBase.cs
@@ -1,5 +1,5 @@
 ﻿using Ciallo.Data;
-using Ciallo.NodeControl;
+using Ciallo.Geometry;
 using Frent;
 
 namespace Ciallo.Tool;
