+++ Ciallo/Tool/PaintFillHover.cs
@@ -1,12 +1,17 @@
-﻿namespace Ciallo.Tool;
+﻿using Ciallo.Rendering;
+using Godot;
+
+namespace Ciallo.Tool;
 
 public class PaintFillHover : HoverBase
 {
     public override void Start()
     {
+        Document.Get<WorldCursorDetectionArea>().MouseDefaultCursorShape = Control.CursorShape.Cross;
     }
 
     public override void End()
     {
+        Document.Get<WorldCursorDetectionArea>().MouseDefaultCursorShape = default;
     }
 }
