+++ Ciallo/Tool/PolylineTransformInteractor.cs
@@ -2,7 +2,6 @@
 using System.Linq;
 using Ciallo.Data;
 using Ciallo.Geometry;
-using Ciallo.NodeControl;
 using Ciallo.Rendering;
 using Frent;
 using Godot;
